Move content helpers into a lib package
Introduce src/jg/hen/lib/ to hold the small, focused content helpers
that were cluttering core.py and living at the top level:

- lib/demo.py: fetch_demo, get_demo_url, is_demo_server_error (from core)
- lib/readme.py: moved from jg.hen.readme

core.py now imports fetch_demo/get_demo_url from lib.demo and keeps only
orchestration. Anchor the packaging .gitignore patterns (/lib/, /lib64/)
to the repo root so they no longer swallow the new source package.
