+++ src/jg/hen/insights/projects.py
@@ -1,7 +1,7 @@
 from typing import Any
 
 from jg.hen.models import RepositoryContext
-from jg.hen.readme import extract_image_urls, extract_title, make_urls_absolute
+from jg.hen.lib.readme import extract_image_urls, extract_title, make_urls_absolute
 from jg.hen.signals import insight, on_repos
