+++ src/jg/hen/rules/has_pinned_repo_with_images.py
@@ -1,5 +1,5 @@
 from jg.hen.models import RepositoryContext, Status
-from jg.hen.readme import extract_image_urls
+from jg.hen.lib.readme import extract_image_urls
 from jg.hen.signals import RuleResult, on_repo, rule
